requirements.txt
aiohttp==3.6.2
async-timeout==3.0.1
attrs==19.3.0
Automat==20.2.0
cffi==1.14.0
chardet==3.0.4
constantly==15.1.0
cryptography==2.8
cssselect==1.1.0
decorator==4.4.2
discord.py==1.3.2
hyperlink==19.0.0
idna==2.9
incremental==17.5.0
lxml==4.5.0
multidict==4.7.5
parsel==1.5.2
Protego==0.1.16
public==2019.4.13
pyasn1==0.4.8
pyasn1-modules==0.2.8
pycparser==2.20
PyDispatcher==2.0.5
PyHamcrest==2.0.2
pyOpenSSL==19.1.0
python-a2s==1.1.4
queuelib==1.5.0
service-identity==18.1.0
six==1.14.0
Twisted==20.3.0
w3lib==1.21.0
websockets==8.1
yarl==1.4.2
zope.interface==5.0.1
